feat: support windows for 2fa tokens
When using 2fa enabled authentication on direct
templates we run into an issue where a 2fa token
has been attached to a field but it's submitted at
a later point. To better facilitate this we have
introduced the ability to have a window of valid tokens.

This won't affect other signing methods since tokens are
verified immediately after they're entered.