feat: automatically sign fields in large documents
Adds a dialog that will display when a certain field threshold
is reached asking the user if they would like to sign non-critical
fields such as name, date, initials, and email with information
that is already available.

Will not prompt the user if there is action authentication on the
document.