feat: add controls for sending completion emails to document owners
Adds a new `ownerDocumentCompleted` to the email settings that
controls whether a document will be sent to the owner upon completion.

This was previously the only email you couldn't disable and didn't
account for users integrating with just the API and Webhooks.

Also adds a flag to the public `sendDocument` endpoint which will
adjust this setting while sendint the document for users who aren't
using `emailSettings` on the `createDocument` endpoint.