fix: element visible race condition
On larger documents we could accidentally start trying to render fields
while not all pages of the PDF have loaded due to us checking for a single
page existing. This would cause an error to be thrown, hard locking those
documents.

This change resolves this by grabbing the highest page number from the given
fields and using it for the visibility check instead.