fix: resolve issue with sealing task on inngest
Currently on inngest the sealing task fails during
decoration stating that it can not find the step "xxx"

My running theory is that this was due to it being a
Promise.all(map(...)) even though that isn't explicitly
disallowed.

This change turns it into a for loop collecting promises
to be awaited after the fact.

Local inngest testing looks promising.